Reorganize utils and services
